FullCarDTOResponse code cleanup PricePerDayDto code cleanup

+++ src/main/java/com/telran/ilcarro/model/web/PricePerDayDto.java
@@ -9,7 +9,6 @@ import lombok.*;
 @Getter
 @Setter
 public class PricePerDayDto {
-
     private String currency;
-    private Double value;
+    private float value;
 }
